% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asyApply.R
\name{asyApply}
\alias{asyApply}
\title{apply a function to each of the assays in an SE, efficiently}
\usage{
asyApply(x, fn, ..., parallel = FALSE)
}
\arguments{
\item{x}{anything descended from a SummarizedExperiment}

\item{fn}{the actual function to run on each assay matrix}
}
\value{
a list of assays having been transformed
}
\description{
apply a function to each of the assays in an SE, efficiently
}
