% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byExtremality.R
\name{byExtremality}
\alias{byExtremality}
\title{For methylation fractions, particularly summarized across entire regions,
we can do better (a lot better) than SD or MAD.  Since we know that}
\usage{
byExtremality(x, k = 500)
}
\arguments{
\item{x}{a matrix of beta values}

\item{k}{how many rows to return (500)}
}
\value{
a smaller matrix (usually) than x, ranked by extremality
}
\description{
max(SD(X_j)) if X_j ~ Beta(a, b) < max(SD(X_j)) if X_j ~ Bernoulli(a/(a+b))
}
\details{
for X having a known mean and SD, hence solvable for a + b by MoM, define

extremality = sd(X_j) / bernoulliSD(mean(X_j))

This function selects the k most extremal rows of x and returns their values.
The resulting matrix tends to be (nearly) optimal for separating clusters.
}
