% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collapseAtDMRs.R
\name{collapseAtDMRs}
\alias{collapseAtDMRs}
\title{Collapse a SummarizedExperiment-like object over a GRanges of DMRs, sensibly.
Handy for things like model testing, DMR plotting, GenometriCorr, lars, etc.}
\usage{
collapseAtDMRs(x, y, how = c("median", "mean", "sum", "max", "min"),
  impute = TRUE)
}
\arguments{
\item{x}{anything descended from a SummarizedExperiment}

\item{y}{differentially whatevered regions (DMRs) as a GRanges}

\item{how}{how to summarize the assay values (default: median)}

\item{impute}{should NA entries in input matrices be imputed? (yes)}
}
\value{
an object with same colData but with new rowRanges & assays
}
\description{
FIXME: let the user provide their own summarizer function
}

