% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseMatrix.R
\name{collapseMatrix}
\alias{collapseMatrix}
\title{collapse variable, and perhaps overlapping, rows of a matrix-like structure}
\usage{
collapseMatrix(x, z, fn, impute = TRUE)
}
\arguments{
\item{x}{anything descended from a SummarizedExperiment}

\item{z}{the listed matrix indices within DMRs (NOT THE ACTUAL DMRs!)}

\item{f}{the name of the function to be used to summarize x over z}
}
\value{
a list of assays, having been transformed
}
\description{
collapse variable, and perhaps overlapping, rows of a matrix-like structure
}

