% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapseOverRegions.R
\name{collapseOverRegions}
\alias{collapseOverRegions}
\title{alias to collapseAtDMRs}
\usage{
collapseOverRegions(x, regions, how = c("median", "mean", "sum", "max",
  "min"), impute = TRUE)
}
\arguments{
\item{x}{the SummarizedExperiment-like thing to collapse}

\item{regions}{the regions to collapse over, as a GRanges}

\item{how}{the function used to collapse x over the regions (median)}

\item{impute}{should missing values be imputed first? (TRUE)}
}
\value{
an object not unlike x, but collapsed over the regions.
}
\description{
alias to collapseAtDMRs
}
