\name{DMRcate-internal}
\title{
Internal DMRcate objects and functions
}
\alias{annot}
\alias{dmrcate.output}
\alias{extractCoords}
\alias{fitParallel}
\alias{KernelSums}
\alias{KernelTest}
\alias{Segment}
\alias{SparseDeltas}
\description{
Internal DMRcate objects and functions
}
\keyword{internal}

