% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCoords.R
\name{extractCoords}
\alias{extractCoords}
\title{Extract coordinates (for transformation into GRanges; somewhat superfluous)
Thank you to Xavier Pastor from Bioconductor mailing list for this patch}
\usage{
extractCoords(xx)
}
\arguments{
\item{xx}{a set of strings in the format chr:start-end[:strand]}
}
\value{
a data.frame with columns chrom, chromStart, chromEnd
}
\description{
Extract coordinates (for transformation into GRanges; somewhat superfluous)
Thank you to Xavier Pastor from Bioconductor mailing list for this patch
}
