% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractRanges.R
\name{extractRanges}
\alias{extractRanges}
\title{extract GRanges corresponding to DMRs from the results of a dmrcate run}
\usage{
extractRanges(dmrcated, bySign = FALSE, atCutoff = NULL, ...)
}
\arguments{
\item{dmrcated}{an object of class dmrcate.output}

\item{bySign}{boolean, shall the DMRs be extracted by sign?  (FALSE)}

\item{atCutoff}{a p-value cutoff desired for extraction, or NULL (NULL)}
}
\description{
extract GRanges corresponding to DMRs from the results of a dmrcate run
}
