% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDMRbetas.R
\name{getDMRbetas}
\alias{getDMRbetas}
\title{extract disjoint GRanges with methylation fraction (beta) diffs as scores
the result can be usefully exported as a bed, bedGraph, or bigWig file}
\usage{
getDMRbetas(dmrcated, minDiff = 0.1, bySign = FALSE, withDMLs = FALSE,
  ...)
}
\arguments{
\item{dmrcated}{an object of class dmrcate.output}

\item{minDiff}{get rid of DMRs with less than this beta difference (0.1)}

\item{bySign}{split the results into a GRangesList by sign? (FALSE)}

\item{withDMLs}{return both DMRs and DMLs? (Forces bySign to FALSE)}
}
\value{
a GRanges or GRangesList, depending values of bySign and withDMLs
}
\description{
extract disjoint GRanges with methylation fraction (beta) diffs as scores
the result can be usefully exported as a bed, bedGraph, or bigWig file
}
