% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDMRdepth.R
\name{getDMRdepth}
\alias{getDMRdepth}
\title{function to allow "piling up" differentially methylated region significance}
\usage{
getDMRdepth(output, stepSize = 1, bridgeSize = 2000, bridgeP = 0.01, ...)
}
\arguments{
\item{output}{DMRcate output (of class "dmrcate.output" as it happens)}

\item{stepSize}{multiplier for scores. Typically if using 10**-y, it's 1}

\item{bridgeSize}{bridge regions between significant DMRs up to this size}

\item{bridgeP}{don't bridge gaps with adjusted p-values less than this}
}
\description{
function to allow "piling up" differentially methylated region significance
}

