% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDMRdiffs.R
\name{getDMRdiffs}
\alias{getDMRdiffs}
\title{Create GRanges of imputed beta FC between samples to dump to a BigWig}
\usage{
getDMRdiffs(output, ...)
}
\arguments{
\item{output}{DMRcate output (of class "dmrcate.output", as it happens)}
}
\value{
summarized unscaled effect size estimates for the output
}
\description{
Create GRanges of imputed beta FC between samples to dump to a BigWig
}

