% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDMRsAndVMRs.R
\name{getDMRsAndVMRs}
\alias{getDMRsAndVMRs}
\title{get both DMRs and VMRs from the same dataset}
\usage{
getDMRsAndVMRs(x, design = NULL, contrasts = F, cont.matrix = NULL,
  coef = 2, ...)
}
\arguments{
\item{x}{matrix of M-values or SummarizedExperiment-derived object}

\item{design}{design matrix (for limma) or Surv object/matrix (for Cox)}

\item{contrasts}{apply contrasts to e.g. paired samples? (only for limma)}

\item{cont.matrix}{if contrasts=T, supply a contrast matrix here}

\item{coef}{column of the design matrix to fit DMRs (or "all")?}
}
\value{
an object of type dmrcate.output
}
\description{
get both DMRs and VMRs from the same dataset
}

