% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSafeProbes.R
\name{getSafeProbes}
\alias{getSafeProbes}
\title{get "safe" probes based on EPIC remappings}
\usage{
getSafeProbes(object, rmXY = FALSE, ...)
}
\arguments{
\item{object}{a matrix of M or Beta values}

\item{rmXY}{remove chrX and chrY? (FALSE)}
}
\value{
the same matrix minus "unsafe" probes
}
\description{
see Zhou et al., https://doi.org/10.1093/nar/gkw967
}
