% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSimpleMethHeatMap.R
\name{getSimpleMethHeatMap}
\alias{getSimpleMethHeatMap}
\title{pretty much what it says, except that (ironically) it calls ComplexHeatmap}
\usage{
getSimpleMethHeatMap(x, k = 100, asSNPs = F, binary = F, rotate = F,
  clustering_method_rows = "ward", clustering_method_columns = "complete",
  clustering_distance_rows = ifelse(binary, "binary", "manhattan"),
  clustering_distance_columns = ifelse(binary, "binary", "manhattan"), ...)
}
\arguments{
\item{x}{a matrix, GenomicRatioSet, or other rectangular object}

\item{k}{choose the top [this many] features by SD}
}
\value{
invisibly, the same thing as Heatmap(...)
}
\description{
pretty much what it says, except that (ironically) it calls ComplexHeatmap
}
