% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDMRsAndVMRs.R
\name{getVMRs}
\alias{getVMRs}
\title{get variably methylated regions}
\usage{
getVMRs(x, pcutoff = 0.1, ...)
}
\arguments{
\item{x}{matrix of M-values or SummarizedExperiment-derived object}

\item{pcutoff}{p-value cutoff; if not specified, step from 10**-1:10**-8}
}
\value{
an object of type dmrcate.output
}
\description{
get variably methylated regions
}
