% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepM.R
\name{prepM}
\alias{prepM}
\title{this is the core of prepMvals but omits sanity checking; useful if collapsed}
\usage{
prepM(M, cutoff = 10)
}
\arguments{
\item{M}{M-value matrix}

\item{cutoff}{extremes at which to truncate M-values}
}
\value{
a matrix
}
\description{
this is the core of prepMvals but omits sanity checking; useful if collapsed
}

