% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepMvals.R
\name{prepMvals}
\alias{prepMvals}
\title{this is the recommended user-facing probe-name version}
\usage{
prepMvals(grset, cutoff = 10, returnBetas = FALSE)
}
\arguments{
\item{grset}{A GenomicRatioSet or something very much like it}

\item{cutoff}{Maximum absolute M-value before truncation (10)}

\item{returnBetas}{Return beta values instead of M-values (FALSE)}
}
\value{
A matrix of tidied and truncated M- or beta-values
}
\description{
this is the recommended user-facing probe-name version
}

