% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dmrcate.output.R
\name{print.dmrcate.output}
\alias{print.dmrcate.output}
\title{make it harder for me to shoot myself in the foot calling `some_DMR_results`}
\usage{
\method{print}{dmrcate.output}(object)
}
\arguments{
\item{an}{object of class dmrcate.output}
}
\description{
make it harder for me to shoot myself in the foot calling `some_DMR_results`
}
