% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDMRdepth.R
\title{getDMRdepth}
\name{getDMRdepth}
\alias{getDMRdepth}
\usage{
getDMRdepth(output, stepSize = 1, ...)
}
\arguments{
\item{output}{DMRcate output (of class "dmrcate.output", as it happens)}

\item{stepSize}{multiplier for scores. Typically if using 10**-y, leave at 1}
}

