% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDMRsAndVMRs.R
\name{getDMRs}
\alias{getDMRs}
\title{omnibus DMR finder starting from a m-values and a design/Surv matrix}
\usage{
getDMRs(x, design = NULL, contrasts = FALSE, cont.matrix = NULL,
  coef = 2, pcutoff = NULL, betacutoff = 0.1, keep.raw.DMLs = TRUE,
  p.adjust.method = "limma", what = c("limma", "cox"), ...)
}
\arguments{
\item{x}{matrix of M-values or SummarizedExperiment-derived object}

\item{design}{design matrix (for limma) or Surv object/matrix (for Cox)}

\item{contrasts}{apply contrasts to e.g. paired samples? (only for limma)}

\item{cont.matrix}{if contrasts=T, supply a contrast matrix here}

\item{coef}{column of the design matrix to fit DMRs (or "all")?}

\item{pcutoff}{p-value cutoff; if not specified, step from 10**-1:10**-8}

\item{betacutoff}{DMRs must have at least this great a maxbetaFC difference}

\item{p.adjust.method}{how to control the FDR, default is limma-style}

\item{what}{(not used yet) whether to use limma or Cox PH to tag DMRs}
}
\value{
an object of type dmrcate.output
}
\description{
omnibus DMR finder starting from a m-values and a design/Surv matrix
}

